Webpack fixes, contract fixes
I changed the string variables in the Movie struct to bytes32 variables. This reduces gas costs and makes it less likely that you will run out of gas when adding a new movie.

I added an updateMovie() function to allow movie information to be updated after it has been added.

I added a paused flag to keep track of whether purchases are currently paused. This allows the contract owner to temporarily disable purchases if needed.

I added a check to make sure purchases are not currently paused before allowing a purchase to be made.

I added a check to make sure the movie ID passed to getMovie() is valid.

I added a check to make sure the movie ID passed to buyMovie() is valid.

And refactored your webpack now working